minor review fixes
